feat: Refactor RecipeManager into static utility class and rename PrepareRecipesEvent to PrepareRecipesListener
