+++ src/main/java/zone/vao/nexoAddon/handlers/RecipeManager.java
@@ -1,6 +1,8 @@
 package zone.vao.nexoAddon.handlers;
 
 import com.nexomc.nexo.api.NexoItems;
+import lombok.Getter;
+import lombok.Setter;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
@@ -15,79 +17,21 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.util.*;
 
+
 public class RecipeManager {
-    private final List<NamespacedKey> registeredRecipes = new ArrayList<>();
-    private final Map<NamespacedKey, String> recipeConfigMap = new HashMap<>();
-    private File recipeFile;
-    private FileConfiguration recipeConfig;
-
-    public RecipeManager() {
-        loadRecipeFile();
-    }
-
-    private void loadRecipeFile() {
-        File recipeFolder = new File(NexoAddon.getInstance().getDataFolder(), "recipes");
-        if (!recipeFolder.exists()) {
-            recipeFolder.mkdirs();
-        }
-
-        recipeFile = new File(recipeFolder, "smithing.yml");
-
-        if (!recipeFile.exists()) {
-            try {
-                InputStream resourceStream = NexoAddon.getInstance().getResource("recipes/smithing.yml");
-                if (resourceStream == null) {
-                    return;
-                }
-
-                Files.copy(resourceStream, recipeFile.toPath());
-            } catch (IOException e) {
-                Bukkit.getLogger().severe("Failed to generate smithing.yml: " + e.getMessage());
-            }
-        }
-
-        recipeConfig = YamlConfiguration.loadConfiguration(recipeFile);
-    }
-
-    public void loadRecipes() {
-        Bukkit.getScheduler().runTask(NexoAddon.getInstance(), () -> {
-            clearRegisteredRecipes();
-
-            if (recipeConfig == null) {
-                Bukkit.getLogger().severe("Recipes file not found");
-                return;
-            }
-
-            recipeConfig.getKeys(false).forEach(recipeId -> {
-                try {
-                    addSmithingTransformRecipe(recipeId, recipeConfig);
-                } catch (IllegalArgumentException e) {
-                    Bukkit.getLogger().warning("Failed to load recipe " + recipeId + ": " + e.getMessage());
-                }
-            });
-        });
-    }
+    @Getter
+    private static final List<NamespacedKey> registeredRecipes = new ArrayList<>();
+    @Getter
+    private static final Map<NamespacedKey, String> recipeConfigMap = new HashMap<>();
+    @Getter
+    @Setter
+    private static File recipeFile;
+    @Getter
+    @Setter
+    private static FileConfiguration recipeConfig;
 
 
-    public Map<NamespacedKey, String> getRecipeConfigMap() {
-        return recipeConfigMap;
-    }
-
-    public FileConfiguration getRecipeConfig() {
-        return recipeConfig;
-    }
-
-
-    private void clearRegisteredRecipes() {
-        registeredRecipes.forEach(key -> {
-            Bukkit.removeRecipe(key);
-            Bukkit.getLogger().info("Removed recipe: " + key);
-        });
-        registeredRecipes.clear();
-        recipeConfigMap.clear();
-    }
-
-    private void addSmithingTransformRecipe(String recipeId, FileConfiguration config) {
+    public static void addSmithingTransformRecipe(String recipeId, FileConfiguration config) {
         ItemStack resultTemplate = parseItem(config, recipeId + ".result");
         RecipeChoice template = parseRecipeChoice(config, recipeId + ".template");
         RecipeChoice base = parseRecipeChoice(config, recipeId + ".base");
@@ -110,7 +54,7 @@ public class RecipeManager {
         }
     }
 
-    private ItemStack parseItem(FileConfiguration config, String path) {
+    private static ItemStack parseItem(FileConfiguration config, String path) {
         String nexoItemId = config.getString(path + ".nexo_item");
         if (nexoItemId != null) return Objects.requireNonNull(NexoItems.itemFromId(nexoItemId)).build();
 
@@ -120,9 +64,10 @@ public class RecipeManager {
         return material != null ? new ItemStack(material) : null;
     }
 
-    private RecipeChoice parseRecipeChoice(FileConfiguration config, String path) {
+    private static RecipeChoice parseRecipeChoice(FileConfiguration config, String path) {
         String nexoItemId = config.getString(path + ".nexo_item");
-        if (nexoItemId != null) return new RecipeChoice.ExactChoice(Objects.requireNonNull(NexoItems.itemFromId(nexoItemId)).build());
+        if (nexoItemId != null && NexoItems.itemFromId(nexoItemId) != null)
+            return new RecipeChoice.ExactChoice(NexoItems.itemFromId(nexoItemId).build());
 
         String materialName = config.getString(path + ".minecraft_item");
         assert materialName != null;
